Small refactor update password page
